fix(drizzle): fix drizzle config and scripts
