+++ packages/api/src/drizzle/schema.ts
@@ -1,1 +1,1 @@
-export * from "../users/user.schema.js";
+export * from "../users/user.schema";
